# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Recipes",
    "description": "Recipe Share Site API",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api-docs": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/users/": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "412": {
            "description": "Precondition Failed"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "412": {
            "description": "Precondition Failed"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/recipes/": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "recipe_name": {
                  "example": "any"
                },
                "recipe_instructions": {
                  "example": "any"
                },
                "ingredient1": {
                  "example": "any"
                },
                "ingredient2": {
                  "example": "any"
                },
                "ingredient3": {
                  "example": "any"
                },
                "ingredient4": {
                  "example": "any"
                },
                "ingredient5": {
                  "example": "any"
                },
                "ingredient6": {
                  "example": "any"
                },
                "ingredient7": {
                  "example": "any"
                },
                "ingredient8": {
                  "example": "any"
                },
                "ingredient9": {
                  "example": "any"
                },
                "ingredient10": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "412": {
            "description": "Precondition Failed"
          }
        }
      }
    },
    "/recipes/{id}": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "recipe_name": {
                  "example": "any"
                },
                "recipe_instructions": {
                  "example": "any"
                },
                "ingredient1": {
                  "example": "any"
                },
                "ingredient2": {
                  "example": "any"
                },
                "ingredient3": {
                  "example": "any"
                },
                "ingredient4": {
                  "example": "any"
                },
                "ingredient5": {
                  "example": "any"
                },
                "ingredient6": {
                  "example": "any"
                },
                "ingredient7": {
                  "example": "any"
                },
                "ingredient8": {
                  "example": "any"
                },
                "ingredient9": {
                  "example": "any"
                },
                "ingredient10": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "412": {
            "description": "Precondition Failed"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    }
  }
}